refactor(api): consolidate Cypher validation and rewriting into cypher_sanitizer module
